This PR contains two commits which speed up pre- and post-processing. The net effect of the two changes is to bring `CARP(presidential_speech)` down to about 0.9s and `CBASS(presidential_speech)` to 1.4s. 

I will copy the descriptions from the original PR that proposed them (#22), updating a few references:

----- 

1c6d7f9 makes two minor speed-ups to pre-processing:

    Avoid Reduce(, lapply(...)) in favor of outer
    (which is fully vectorized) to construct edge
    incidence matrices
    Construct Vmat directly rather than in a loop

I'm not sure what these functions are exactly supposed to be calculating, this change passes the `all.equal` test on the following (and save about a second on total time).

```
list(
  CARPVIZ   = CARP(presidential_speech, alg.type="carpviz"), 
  CARP1.2   = CARP(presidential_speech, alg.type="carp", t=1.2), 
  CARP1.1   = CARP(presidential_speech, alg.type="carp", t=1.1), 
  CARP1.05  = CARP(presidential_speech, alg.type="carp", t=1.05), 
  CBASSVIZ  = CBASS(presidential_speech, alg.type="cbassviz"), 
  CBASS1.2  = CBASS(presidential_speech, alg.type="cbass", t=1.2), 
  CBASS1.1  = CBASS(presidential_speech, alg.type="cbass", t=1.1), 
  CBASS1.05 = CBASS(presidential_speech, alg.type="cbass", t=1.05)
)
```

Tangentially related: Running `profvis` on the above, it looks like the last obvious bottle-neck is ISP, which takes 6690ms out of 9970ms (or about 67% of total run time) on the above. (Results from a single benchmark, so might not be robust). In particular, most of that time (~4900ms) is in one step [1]. I'll take a look and see if it can be improved, but on first glance I can't really tell what it's doing.

[1] https://github.com/DataSlingers/clustRviz/blob/master/R/utils.R#L350-L357

---

f7ecb35 speeds up ISP with a one line tweak. After this test, the 8 runs about take about 7 seconds, with the total time spent in ISP being reduced to about 2010ms.

As a result, `CARP(presidential_speech)` now takes about 0.9s while `CBASS(presidential_speech)` is taking about 1.5s.

As before, the `all.equal` test on 8 different versions runs on `presidential_speech` passes (gets the same result as `develop`.)